* Added a cancel feature that can stop file writing at any point in the * process
* This feature required a Controller class to actually manage the worker
* thread and all objects in the application. It manages the connections
* and lifetime of objects. This class kills and creates the XorLogic
* object and worker thread as needed when stopping is requested.

* Remade the object auto-naming/scanning to work properly.

+++ source/xorlogic.cpp
@@ -5,27 +5,24 @@ XorLogic::XorLogic()
     : saveDir(new QDir("../"))
     , openDir(new QDir("../"))
     , overwriteMode(false)
-    , filterList(new QStringList())
     , deleteInput(false)
-    , interrupt(false)
+    , timerMode(false)
 {
+    initTimer();
 }
 
 XorLogic::~XorLogic()
 {
     delete saveDir;
     delete openDir;
-    delete timer;
-    delete filterList;
 }
 
 void XorLogic::initTimer()
 {
     Q_ASSERT(QThread::currentThread() == this->thread());
-    timer = new QTimer(this);
-    timer->setSingleShot(true);
-    connect(timer, &QTimer::timeout, this, &XorLogic::slot_timeout);
-    connect(this, &XorLogic::signal_timeout_start, this, &XorLogic::slot_startButton_clicked);
+    timer.setSingleShot(true);
+    connect(&timer, &QTimer::timeout, this, &XorLogic::timeout);
+    connect(this, &XorLogic::timeout_start, this, &XorLogic::startButton_clicked);
 }
 
 quint32 XorLogic::countFiles()
@@ -53,13 +50,21 @@ QFile* XorLogic::createFile(const QFileInfo &fileInfo, bool overwrite)
 {
     QFile *createdFile;
     QString suffix = "_output";
+    QRegularExpression regexp("(\\d+)");
+    QRegularExpressionMatch match;
     if(!overwrite)
     {
+        match = regexp.match(fileInfo.baseName());
         QString newFilename = saveDir->absolutePath().append("/");
         newFilename.append(fileInfo.baseName());
-        newFilename.append(suffix);
-        qDebug() << "countFiles() returned:" << countFiles();
+        match = regexp.match(newFilename);
+        if(newFilename.contains("_output"))
+            newFilename.remove(match.capturedStart(), match.captured().size());
+        else newFilename.append(suffix);
+
         newFilename.append(QString::number(countFiles() + 1));
+
+        qDebug() << "countFiles() returned:" << countFiles();
         newFilename.append(".txt");
         createdFile = new QFile(newFilename);
         qDebug() << "Overwrite mode: OFF; filepath:" << newFilename;
@@ -85,10 +90,12 @@ uint64_t XorLogic::invertBinary(uint64_t num)
 
 void XorLogic::setupFile()
 {
+    bool skip = false;
     QFileInfoList files = scanForFiles();
+    QList<QFileInfo> deletedFiles;
     for(int i = 0; i < files.size(); i++)
     {
-        emit signal_xor_started(files.at(i).fileName());
+        emit xor_started(files.at(i).fileName());
         qDebug() << "Performing XOR on file: " << files.at(i).fileName();
         QFile *file = createFile(files.at(i), true);
         QFile *outFile = createFile(files.at(i), false);
@@ -99,25 +106,43 @@ void XorLogic::setupFile()
         if(overwriteMode)
         {
             writeFile(file, buffer);
-            emit signal_xor_finished(files.at(i).fileName());
+            emit xor_finished(files.at(i).fileName());
         }
         else if(!overwriteMode)
         {
             if(deleteInput)
             {
+                if(!deletedFiles.empty())
+                {
+                    for(QFileInfo info : deletedFiles)
+                    {
+                        if(file->fileName() == info.fileName())
+                        {
+                            skip = true;
+                            break;
+                        }
+                    }
+                }
+                if(skip) break;
                 deleteFile(file);
+                deletedFiles.append(*file);
             }
             writeFile(outFile, buffer);
-            emit signal_xor_finished(files.at(i).fileName());
+            emit xor_finished(files.at(i).fileName());
         }
-            delete file;
-            delete buffer;
-            delete outFile;
+        skip = false;
+        delete file;
+        delete buffer;
+        delete outFile;
     }
 }
 
 void XorLogic::run()
 {
+    if(QThread::currentThread()->isInterruptionRequested())
+    {
+        return;
+    }
     setupFile();
 }
 
@@ -151,7 +176,7 @@ bool XorLogic::writeFile(QFile *file, QBuffer *buffer)
 {
     QByteArray bytearr(READ_SIZE, '\0');
     quint64 maxProgress = buffer->size();
-    quint64 progress = 0;
+    quint64 progressed = 0;
     if(file->open(QIODevice::WriteOnly))
     {
         if(buffer->open(QIODevice::ReadOnly))
@@ -159,13 +184,14 @@ bool XorLogic::writeFile(QFile *file, QBuffer *buffer)
 
             while(!buffer->atEnd())
             {
-                if(interrupt)
-                    break;
+                if(QThread::currentThread()->isInterruptionRequested()) // при запрошенном прерывании прекращаем работу
+                    return false;
+
                 bytearr = buffer->read(READ_SIZE);
                 // qDebug() << "Written to byte array from buffer:" << bytearr.size();
-                progress += bytearr.size();
+                progressed += bytearr.size();
 
-                emit signal_progress(progress, maxProgress);
+                emit progress(progressed, maxProgress);
 
                 // если нужно дополнение до кратного восьми размера
                 // if(bytearr.size() < 8)
@@ -203,67 +229,65 @@ void XorLogic::performXOR(QByteArray& bytearr)
     }
 }
 
-void XorLogic::slot_btn_openPath_clicked(QString openPath)
+void XorLogic::btn_openPath_clicked(QString openPath)
 {
     openDir->cd(openPath);
     scanForFiles();
     qDebug() << "Button open path clicked, open dir: " << openPath;
 }
 
-void XorLogic::slot_btn_savePath_clicked(QString savePath)
+void XorLogic::btn_savePath_clicked(QString savePath)
 {
     saveDir->cd(savePath);
     qDebug() << "Button save path clicked, save dir:" << savePath;
 }
 
-void XorLogic::slot_modifierEdited(uint64_t mod)
+void XorLogic::modifierEdited(uint64_t mod)
 {
     modifier = mod;
     qDebug() << "Saved modifier:" << modifier;
 }
 
-void XorLogic::slot_overwriteModeChecked(bool isChecked)
+void XorLogic::overwriteModeChecked(bool isChecked)
 {
     overwriteMode = isChecked;
     qDebug() << "Overwrite mode:" << overwriteMode;
 }
 
-void XorLogic::slot_startButton_clicked(uint64_t mod)
+void XorLogic::startButton_clicked()
 {
-    interrupt = false;
-    if(timerMode) startTimer();
+    if(timerMode)
+        startTimer();
     scanForFiles();
-    modifier = mod;
     run();
 }
 
-void XorLogic::slot_btn_stop_clicked()
+void XorLogic::btn_stop_clicked()
 {
-    timer->stop();
-    interrupt = true;
+    timer.stop();
     qDebug() << "Stop button clicked";
 }
 
-void XorLogic::slot_timerModeChecked(bool isChecked)
+void XorLogic::timerModeChecked(bool isChecked)
 {
     timerMode = isChecked;
-    if(!timerMode) timer->stop();
+    if(!timerMode) timer.stop();
     qDebug() << "Timer mode:" << timerMode;
 }
 
 void XorLogic::startTimer()
 {
     quint64 intervalMs = timerInterval * 1000;
-    timer->start(intervalMs);
+    timer.start(intervalMs);
     qDebug() << "Timer started!";
 }
 
-void XorLogic::slot_timeout()
+void XorLogic::timeout()
 {
-    if(!timerMode) timer->stop();
+    if(!timerMode) timer.stop();
     if(busy)
     {
-        timer->stop();
+        timer.stop();
         return;
     }
     qDebug() << "Timeout reached";
@@ -287,38 +311,38 @@ QFileInfoList XorLogic::scanForFiles()
         }
 
         QString extension = parts.last();
-        for(int j = 0; j < filterList->size(); j++)
+        for(int j = 0; j < filterList.size(); j++)
         {
-            if(QString(filterList->at(j)).contains(extension))
+            if(QString(filterList.at(j)).contains(extension))
             {
                 outputList.push_back(files.at(i));
-                qDebug() << "filterList[" << j << "] =" << filterList->at(j);
-                qDebug() << "Contains <<" << extension << ">>?:" << QString(filterList->at(j)).contains(extension);
+                qDebug() << "filterList[" << j << "] =" << filterList.at(j);
+                qDebug() << "Contains <<" << extension << ">>?:" << QString(filterList.at(j)).contains(extension);
                 break;
             }
         }
     }
-    emit signal_scanComplete(outputList);
+    emit scanComplete(outputList);
     return outputList;
 }
 
-void XorLogic::slot_timerMode_time_changed(quint32 time)
+void XorLogic::timerMode_time_changed(quint32 time)
 {
     timerInterval = time;
     qDebug() << "Set timer period to" << timerInterval;
 }
 
-void XorLogic::slot_deleteInputChecked(bool isChecked)
+void XorLogic::deleteInputChecked(bool isChecked)
 {
     deleteInput = isChecked;
 }
 
-void XorLogic::slot_maskChanged(QString mask)
+void XorLogic::maskChanged(QString mask)
 {
     qDebug() << "Mask list changed:" << mask;
-    *filterList = QStringList(mask.split(' ', Qt::SkipEmptyParts));
-    for(int i = 0; i < filterList->size(); i++)
+    filterList = QStringList(mask.split(' ', Qt::SkipEmptyParts));
+    for(int i = 0; i < filterList.size(); i++)
     {
-        qDebug() << "filterList[" << i << "] = " << filterList->at(i);
+        qDebug() << "filterList[" << i << "] = " << filterList.at(i);
     }
 }
